Express data collection as transformations
Previously, the components that were overriding configuration values and
adding gathering build information to add to the build scan were
building intermediate mutable data structures. The content of the
mutable data structures was eventually added back to the Develocity
configuration.

With this patch, the collection of data is expressed as a series of
transformation on the Develocity configuration instead.