Abstract over environment
Previously, the plugin was looking directly at the environment variables
and system properties with no abstraction layer in between. With this
patch, `Env` can be used to access the environment variables and system
properties instead. In tests, a custom `Env` can be passed, so that we
have better control over the environment.