Collect git data in CI only
Previously, the plugin would always gather data from Git. When
running outside of CI, this data is likely to quickly get outdated
between commands, especially when running in interactive mode.

Unfortunately, there's no easy way to determine whether sbt will
run in interactive mode when evaluating the settings to build the
initial state. Therefore, we decided to check instead whether we're in
CI to gather or not the git data.