Remove `Scala Versions` custom value
We're concerned this custom value could be misleading to users, since
they may wrongly think that a build has actually used all the Scala
versions listed here, where in reality the custom value lists all the
Scala versions that are configured in the build. In other words, the
custom values was not reflecting what Scala versions had been used in a
given build.